Move cell views to separate files
