+++ packages/nimble-components/src/table-column/text/cell-view/index.ts
@@ -1,0 +1,38 @@
+import { DesignSystem } from '@microsoft/fast-foundation';
+import { template } from '../../text-base/cell-view/template';
+import type { TableColumnTextCellRecord, TableColumnTextColumnConfig } from '..';
+import { styles } from '../../base/styles';
+import { TableColumnTextCellViewBase } from '../../text-base/cell-view';
+
+declare global {
+    interface HTMLElementTagNameMap {
+        'nimble-table-column-text-cell-view': TableColumnTextCellView;
+    }
+}
+
+/**
+ * A cell view for displaying strings as text
+ */
+export class TableColumnTextCellView extends TableColumnTextCellViewBase<TableColumnTextCellRecord, TableColumnTextColumnConfig> {
+    public override get text(): string {
+        return this.cellRecord.value!;
+    }
+
+    public override get placeholder(): string {
+        return this.columnConfig.placeholder;
+    }
+
+    public override get shouldUsePlaceholder(): boolean {
+        return typeof this.cellRecord.value !== 'string';
+    }
+}
+
+const textCellView = TableColumnTextCellView.compose({
+    baseName: 'table-column-text-cell-view',
+    template,
+    styles
+});
+DesignSystem.getOrCreate().withPrefix('nimble').register(textCellView());
+export const tableColumnTextCellViewTag = DesignSystem.tagFor(
+    TableColumnTextCellView
+);
